requirements/requirements.neopixel.txt
adafruit-circuitpython-neopixel  # 6.3.0
adafruit-circuitpython-pixelbuf  # 1.1.3
